[squash] move csp-report route bind out of server.js, amend tests to validate trust proxy behaviour
